update paths and logout callback for newest userHome
