+++ lib/navigation/routes.dart
@@ -5,7 +5,7 @@ import '../screens/animal_inventory_screen.dart';
 import '../screens/choose_animal_type_screen.dart';
 import '../screens/landing_screen.dart';
 import '../screens/login_screen.dart';
-import '../screens/user_home_screen_UI.dart';
+import '../screens/user_home_screen.dart';
 
 class RouteNames {
   static final routes = {
